+++ inst/shiny/PLPViewer/server.R
@@ -30,7 +30,7 @@ server <- shiny::shinyServer(function(input, output, session) {
   #print(summaryTable)
   
   # need to remove over columns:
-  output$summaryTable <- DT::renderDataTable(DT::datatable(summaryTable[filterIndex(),!colnames(summaryTable)%in%c('addExposureDaysToStart','addExposureDaysToEnd', 'plpResultLocation', 'plpResultLoad')],
+  output$summaryTable <- DT::renderDataTable(DT::datatable(summaryTable[filterIndex(),!colnames(summaryTable)%in%c('Analysis','addExposureDaysToStart','addExposureDaysToEnd', 'plpResultLocation', 'plpResultLoad')],
                                                            rownames= FALSE, selection = 'single',
                                              extensions = 'Buttons', options = list(
                                                dom = 'Blfrtip' , 
@@ -66,6 +66,15 @@ server <- shiny::shinyServer(function(input, output, session) {
   
   output$modelCovariateInfo <- DT::renderDataTable(data.frame(covariates = nrow(plpResult()$covariateSummary),
                                                               nonZeroCount = sum(plpResult()$covariateSummary$covariateValue!=0)))
+  # Download plpresult
+  output$plpResult <- shiny::downloadHandler(
+    filename = function(){
+      "plpResult.rds"
+    },
+    content = function(file) {
+      saveRDS(plpResult(), file)
+    }
+  )
   
   # Downloadable csv of model ----
   output$downloadData <- shiny::downloadHandler(
@@ -200,11 +209,48 @@ server <- shiny::shinyServer(function(input, output, session) {
     }
   })
   
+  #=======================
+  # cal for different evals
+  
+  output$recalSelect = renderUI({
+    types <- unique(plpResult()$performanceEvaluation$calibrationSummary$Eval)
+    selectInput('recal', 'Type:', types)
+  })
+  
+  output$calTable <- shiny::renderTable({
+    if(is.null(plpResult()$performanceEvaluation)){
+      return(NULL)
+    } else{
+      data <- plpResult()$performanceEvaluation$evaluationStatistics
+      data <- as.data.frame(data)
+      
+      data$Metric <- as.character(data$Metric)
+      data$Value <- as.double(as.character(data$Value))
+      
+      ind <- data$Metric %in% c('CalibrationIntercept', 
+                        'CalibrationSlope',
+                        'CalibrationInLarge',
+                        'Emean',
+                         'E90',
+                         'Emax', 
+                        'correctionFactor',
+                        'adjustGradient',
+                        'adjustIntercept')
+      
+      result <- reshape2::dcast(data[ind,],
+                      Eval ~ Metric, value.var = 'Value')
+      row.names(result) <- NULL
+      result
+      
+    }
+  })
+  
   output$cal <- shiny::renderPlot({
     if(is.null(plpResult()$performanceEvaluation)){
       return(NULL)
     } else{
-      plotSparseCalibration2(plpResult()$performanceEvaluation)
+      plotSparseCalibration2(evaluation = plpResult()$performanceEvaluation, 
+                             type = input$recal)
     }
   })
   
@@ -212,10 +258,100 @@ server <- shiny::shinyServer(function(input, output, session) {
     if(is.null(plpResult()$performanceEvaluation)){
       return(NULL)
     } else{
-      tryCatch(plotDemographicSummary(plpResult()$performanceEvaluation),
+      tryCatch(plotDemographicSummary(evaluation = plpResult()$performanceEvaluation, 
+                                      type = input$recal),
                error= function(cond){return(NULL)})
     }
   })
+  
+  #=======================
+  
+  
+  
+  #=======================
+  # NETBENEFIT
+  
+  output$nbSelect = renderUI({
+    types <- unique(plpResult()$performanceEvaluation$thresholdSummary$Eval)
+    selectInput('nbSelect', 'Type:', types)
+  })
+  
+  output$nbTable <- shiny::renderTable({
+    if(is.null(plpResult()$performanceEvaluation)){
+      return(NULL)
+    } else{
+      result <- extractNetBenefit(performanceEvaluation = plpResult()$performanceEvaluation, 
+                        type=input$nbSelect)
+      unique(result)
+    }
+  })
+  
+  output$nbPlot <- shiny::renderPlot({
+    if(is.null(plpResult()$performanceEvaluation)){
+      return(NULL)
+    } else{
+      result <- extractNetBenefit(performanceEvaluation = plpResult()$performanceEvaluation, 
+                                  type=input$nbSelect)
+      result <- unique(result)
+      plot(result$pt, result$netBenefit)
+    }
+  })
+  
+  #=======================
+  
+  
+  #=======================
+  # validation table and selection
+  validationTable <- shiny::reactive(dplyr::filter(summaryTable[filterIndex(),],
+                                                   Analysis == summaryTable[filterIndex(),'Analysis'][trueRow()]))
+  
+  output$validationTable <- DT::renderDataTable(dplyr::select(validationTable(),c(Analysis, Dev, Val, AUC)), rownames= FALSE)
+  
+  valFilterIndex <- shiny::reactive({getFilter(validationTable(), input)})
+  valSelectedRow <- shiny::reactive({
+    if(is.null(input$validationTable_rows_selected[1])){
+      return(1)
+    }else{
+      # return(input$validationTable_rows_selected[1])
+      return(input$validationTable_rows_selected)
+    }
+  })
+  
+  # plots for the validation section. todo: add the development?
+  
+  valResult <- shiny::reactive({
+    valtemplist <- list()
+    valTable <- validationTable()
+    rows <- sort(valSelectedRow())
+    names <- valTable[rows, "Val"]
+    for (i in 1:length(rows)){
+      valtemplist[[i]] <- getPlpResult(result,validation,valTable, 'file', rows[i])
+    }
+    list(results = valtemplist, databaseName = names)
+  })
+  
+  valPlots <- shiny::reactive({
+    results <- valResult()
+    if(is.null(valResult()$results[[1]]$performanceEvaluation)){
+      return(NULL)
+    } else{
+      valCalPlot <- plotCals(evaluationList = valResult()$results, 
+                             modelNames =  valResult()$databaseName)
+      valRocPlot <- plotRocs(evaluationList = valResult()$results, 
+                             modelNames = valResult()$databaseName)
+      list(valRocPlot= valRocPlot, valCalPlot = valCalPlot)
+    }
+  })
+  
+  output$valRoc <- shiny::renderPlot({
+    try(valPlots()$valRocPlot)
+  })
+  output$valCal <- shiny::renderPlot({
+    try(valPlots()$valCalPlot)
+  })
+  #=======================
+  
+